Refactor: create pagination handler and removed pagination code from controller to use just the fuctionality
